+++ src/view/filter.js
@@ -1,4 +1,4 @@
-import AbstractView from './abstract.js';
+import AbstractView from './abstract';
 
 const createFilterItemTemplate = (filter, isChecked) => {
   const {title, count} = filter;
