+++ src/view/load-more-button.js
@@ -1,4 +1,4 @@
-import AbstractView from './abstract.js';
+import AbstractView from './abstract';
 
 const createLoadMoreButtonTemplate = () => {
   return (
