+++ src/view/site-menu.js
@@ -1,4 +1,4 @@
-import AbstractView from './abstract.js';
+import AbstractView from './abstract';
 
 const createSiteMenuTemplate = () => {
   return (
