+++ src/view/sort.js
@@ -1,4 +1,4 @@
-import AbstractView from './abstract.js';
+import AbstractView from './abstract';
 
 const createSortTemplate = () => {
   return `<div class="board__filter-list">
