+++ src/view/task-list.js
@@ -1,4 +1,4 @@
-import AbstractView from './abstract.js';
+import AbstractView from './abstract';
 
 const createTasksTemplate = () => {
   return `<div class="board__tasks"></div>`;
